Refactoring classes to make a jar
